renames and doc improvements
