+++ extension/public/manifest.json
@@ -16,7 +16,9 @@
     "default_path": "sidepanel.html"
   },
   "externally_connectable": {
-    "matches": ["<all_urls>"]
+    "matches": [
+      "<all_urls>"
+    ]
   },
   "permissions": [
     "sidePanel",
@@ -30,20 +32,14 @@
   "background": {
     "service_worker": "build/background/index.js"
   },
-  "host_permissions": ["<all_urls>"],
+  "host_permissions": [
+    "<all_urls>"
+  ],
   "content_scripts": [
     {
-      "matches": ["<all_urls>"],
-      "exclude_globs": [
-        "https://connect.pilot.gnosisguild.org/",
-        "about:*",
-        "chrome:*"
+      "matches": [
+        "<all_urls>"
       ],
-      "run_at": "document_start",
-      "js": ["build/pilot-connect/contentScripts/connectPilotToDApp.js"]
-    },
-    {
-      "matches": ["<all_urls>"],
       "exclude_globs": [
         "https://connect.pilot.gnosisguild.org/",
         "about:*",
@@ -51,22 +47,42 @@
       ],
       "run_at": "document_start",
       "js": [
-        "build/pilot-connect-monitor/contentScripts/setupConnectionMonitoring.js"
+        "build/connect/contentScripts/dApp.js"
       ]
     },
     {
-      "matches": ["https://connect.pilot.gnosisguild.org/"],
+      "matches": [
+        "https://connect.pilot.gnosisguild.org/"
+      ],
       "run_at": "document_start",
       "all_frames": true,
       "js": [
-        "build/pilot-connect/contentScripts/connectPilotToInjectedIFrame.js"
+        "build/connect/contentScripts/connectIframe.js"
+      ]
+    },
+    {
+      "matches": [
+        "<all_urls>"
+      ],
+      "exclude_globs": [
+        "https://connect.pilot.gnosisguild.org/",
+        "about:*",
+        "chrome:*"
+      ],
+      "run_at": "document_start",
+      "js": [
+        "build/monitor/contentScript/main.js"
       ]
     }
   ],
   "web_accessible_resources": [
     {
-      "resources": ["*"],
-      "matches": ["<all_urls>"]
+      "resources": [
+        "*"
+      ],
+      "matches": [
+        "<all_urls>"
+      ]
     }
   ]
 }
