+++ extension/src/connect/contentScripts/connectIframe.ts
@@ -5,7 +5,5 @@ import { injectScript } from '@/utils'
 if (document.documentElement.dataset.__zodiacPilotInjected !== 'true') {
   document.documentElement.dataset.__zodiacPilotInjected = 'true'
 
-  injectScript(
-    'build/pilot-connect/injectedScripts/connectOriginalWalletToPilot.js'
-  )
+  injectScript('build/inject/injectedScript/main.js')
 }
