+++ extension/src/connect/injectedScript/main.ts
