+++ extension/src/inject/contentScript/main.ts
@@ -24,9 +24,7 @@ if (
   document.documentElement.dataset.__zodiacPilotInjected = 'true'
   document.documentElement.dataset.__zodiacPilotConnected = 'true'
 
-  injectScript(
-    'build/injected-wallet-connect/injectedScripts/enableInjectedProvider.js'
-  )
+  injectScript('build/inject/injectedScript/main.js')
 
   // relay rpc requests from the InjectedProvider in the tab to the Eip1193Provider in the panel
   window.addEventListener(
