+++ extension/src/inject/contentScript/probeChainId.ts
