+++ extension/src/inject/injectedScript/InjectedProvider.ts
