+++ extension/src/inject/injectedScript/announceProvider.ts
