+++ extension/src/inject/injectedScript/ensureInjectedProvider.ts
