+++ extension/src/inject/injectedScript/main.ts
