+++ extension/src/inject/injectedScript/maskMetaMaskAnnouncement.ts
