+++ extension/src/monitor/injectedScript/hint.html
@@ -1,7 +1,7 @@
 <template id="zodiac-template">
   <link
     rel="stylesheet"
-    href="{{BASE_PATH}}build/pilot-connect-monitor/injectedScripts/handleConnectionStatusChange.css"
+    href="{{BASE_PATH}}build/monitor/injectedScript/main.css"
   />
 
   <div
