+++ extension/src/monitor/injectedScript/main.ts
