+++ extension/src/monitor/injectedScript/renderHint.ts
