+++ extension/src/panel/providers/injected-provider/useConnectProvider.tsx
@@ -1,10 +1,10 @@
-import { useWindowId } from '@/bridge'
 import { CHAIN_CURRENCY, CHAIN_NAME, EXPLORER_URL, RPC } from '@/chains'
 import { infoToast } from '@/components'
 import { Eip1193Provider } from '@/types'
 import { MutableRefObject, useEffect, useState } from 'react'
 import toast from 'react-hot-toast'
 import { ChainId } from 'ser-kit'
+import { useWindowId } from '../../../inject/bridge'
 import { ConnectionStatus } from '../connectTypes'
 import { ConnectProvider } from './ConnectProvider'
 import { InjectedWalletError } from './InjectedWalletError'
